Extract villager editing to separate control
